+++ tests/test_search_field.py
@@ -1,0 +1,7 @@
+from wtforms_components import SearchField
+
+from tests import FieldTestCase
+
+
+class TestSearchField(FieldTestCase):
+    field_class = SearchField